pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vidseqpy"
version = "0.1.0"
description = "Python tool for processing video sequences"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "moviepy>=1.0.3,<2.0.0",
    "numpy>=1.21.0,<2.0.0",
    "Pillow>=10.0.0",
    "tqdm>=4.65.0",
    "argcomplete>=3.1.0"
]

[project.urls]
"Homepage" = "https://github.com/haschek/vidseqpy"
"Bug Tracker" = "https://github.com/haschek/vidseqpy/issues"
